feat(floating-button-overhaul): done
